Revert gradle language back to Groovy
This effectively reverts 6fa1abed9bd8eee9cdeb40d72a8b7267562e74ad

Android Studio in the current stable version (3.3) is not able to recognize imported projects that use Gradle Kotlin DSL as Android projects. I am not sure if it is a problem in Android Studio or Kotlin DSL. There is a bug report in the Kotlin DSL repo here:

https://github.com/gradle/kotlin-dsl/issues/1166

Unfortunately none of the workarounds reported above worked for me. The only way to import the project was to:

1) Checkout the latest commit that used the Groovy DSL
2) Import the project in Android Studio
3) Checkout master

This is of course unnaceptable for any potential contributors, therefore for now we are reverting to Groovy DSL. However, we keep the gradle files in the format that allows swift conversion to Kotlin once the problem described above is solved.